test: separate study test class in nextstep-mvc
