+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/JsonViewTest.java
@@ -1,5 +1,6 @@
 package nextstep.mvc.tobe;
 
+import nextstep.mvc.tobe.test.Car;
 import nextstep.mvc.tobe.view.JsonView;
 import nextstep.mvc.tobe.view.View;
 import nextstep.utils.JsonUtils;
