+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/scanner/ControllerScannerTest.java
@@ -1,7 +1,6 @@
 package nextstep.mvc.tobe.scanner;
 
-import nextstep.mvc.tobe.MyController;
-import nextstep.mvc.tobe.scanner.ControllerScanner;
+import nextstep.mvc.tobe.test.MyController;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.InvocationTargetException;
@@ -10,10 +9,11 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ControllerScannerTest {
+    private static final String BASE_PACKAGE = "nextstep.mvc.tobe.test";
 
     @Test
     void scan() throws InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
-        Map<Class<?>, Object> controllerInstanceMap = ControllerScanner.scan("nextstep.mvc.tobe");
+        Map<Class<?>, Object> controllerInstanceMap = ControllerScanner.scan(BASE_PACKAGE);
         assertThat(controllerInstanceMap.size()).isEqualTo(1);
         assertThat(controllerInstanceMap.containsKey(MyController.class)).isTrue();
         assertThat(controllerInstanceMap.get(MyController.class)).isEqualTo(controllerInstanceMap.get(MyController.class));
