+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/test/Car.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.test;
 
 import java.util.Objects;
