+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/test/TestUser.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.test;
 
 public class TestUser {
     private String userId;
