+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/test/TestUserController.java
@@ -1,5 +1,6 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.test;
 
+import nextstep.mvc.tobe.ModelAndView;
 import nextstep.web.annotation.PathVariable;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
