+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/test/User.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.test;
 
 public class User {
     private String userId;
