+++ nextstep-mvc/src/test/java/nextstep/utils/JsonUtilsTest.java
@@ -1,6 +1,6 @@
 package nextstep.utils;
 
-import nextstep.mvc.tobe.Car;
+import nextstep.mvc.tobe.test.Car;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
